backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.3
python-dotenv==1.1.0
psycopg2-binary==2.9.10
google-generativeai==0.8.5
numpy==2.2.1
requests==2.32.3
python-multipart==0.0.20
asyncio==3.4.3